refactor(sim-plugin): store eBPF counters as 32-bit values
`CounterArray` documents its counters as unsigned 32-bit values updated
by a `bit<32>` operand, so the counts are `u32` and wrap at that width
instead of at a host integer width.

+++ p4spec-rust/tests/sim-plugin/ebpf/mod.rs
@@ -22,7 +22,7 @@ fn counter_counts(
     >,
     value_arch: Value,
     names: &[&str],
-) -> Vec<i64> {
+) -> Vec<u32> {
     let values = names
         .iter()
         .map(|name| make::text(runner.arena_mut(), (*name).to_owned(), Span::default()).unwrap())
